Functional update: support 128, 192 and 256 bit
The code was updated to support 128, 192 and 256-bit AES ciphers
and any length of IV.
The test cases were extended to cover all those present in the
NIST GCM revised spec.
The test script was fixed to use "should" correctly.
README was updated to reflect changes.
Some type casts were added to silence warnings.
Added functionality, so version was bumped to 0.2.0.