Fix memory leak bug
I was using the wrong buffer allocation method (CopyBuffer), it did
a copy of the C++ data to the Node Buffer object, but the C++ memory
that was allocated wasn't freed.  Now I use NewBuffer, which takes
control of the C++ buffers and frees them when the Buffer object is
garbage collected.
Also updated dependencies.